Add 2 pictures for the last version.
After one unsuccessful attempt to make a homemade double-layer
PCB, I could make a working homemade double-layer PCB. I got a 2 degree
bonus for that. It was the first time in my life to get marks which is
higher than anyone in the class.

I learned a lot from this homemade double-layer PCB.